Move files to better locations, optimize Geom classes.
