+++ core/src/main/scala/archery/Joined.scala
