+++ core/src/main/scala/archery/Node.scala
