+++ core/src/main/scala/archery/RTree.scala
